add rate limiter

+++ src/routes/dashboard/+page.server.ts
@@ -5,6 +5,7 @@ import { error, type Actions, fail, redirect } from "@sveltejs/kit";
 import { prisma } from "$lib/server/prisma";
 import { deleteTaskSchema } from "$lib/components/dashboard/tasks/task-table.svelte";
 import { auth } from "$lib/server/lucia";
+import { ratelimit } from "$lib/server/redis";
 
 export const load: PageServerLoad = async (event) => {
 	if (!event.locals.session) throw redirect(302, "/login");
@@ -32,6 +33,15 @@ export const actions: Actions = {
 		const form = await superValidate(event.request, createTaskSchema);
 		if (!form.valid) return fail(400, { form });
 
+		const { success, reset } = await ratelimit.create.limit(event.locals.session.user.userId);
+
+		if (!success) {
+			const timeRemaining = Math.floor((reset - Date.now()) / 1000);
+			return message(form, `You are doing that too much, wait ${timeRemaining}s and try again.`, {
+				status: 429
+			});
+		}
+
 		try {
 			await prisma.task.create({
 				data: {
@@ -53,6 +63,15 @@ export const actions: Actions = {
 		const form = await superValidate(event.url, deleteTaskSchema);
 
 		if (!form.valid) return fail(400, { form });
+
+		const { success, reset } = await ratelimit.delete.limit(event.locals.session.user.userId);
+
+		if (!success) {
+			const timeRemaining = Math.floor((reset - Date.now()) / 1000);
+			return message(form, `You are doing that too much, wait ${timeRemaining}s and try again.`, {
+				status: 429
+			});
+		}
 
 		try {
 			await prisma.task.delete({
